Added working CAS entry filter which redirects to the server. Added to TenantDTO.
The CAS Entry Filter at /api/cas/entry will redirect the user to the correct CAS Server with the service as a query parameter. This is now integrated with the welcome guard (still need to make an exception for the admin page).

The TenantDTO now includes the tenant's name and authentication method. Removed /api/tenant/name endpoint as it is now included in /api/tenant.